Do not depend on Android components for LibraryLoader. Rename base module to common.
