+++ annotations/src/main/java/com/mapbox/annotation/module/MapboxModuleType.kt
@@ -1,7 +1,7 @@
 package com.mapbox.annotation.module
 
 /**
- * Desribes Map Module's type. The annotatted implementation class needs to extends the base interface.
+ * Describes Map Module's type. The annotated implementation class needs to extends the base interface.
  * @param interfacePackage package of the base interface.
  * @param interfaceClassName class name of the base interface.
  */
@@ -10,5 +10,5 @@ enum class MapboxModuleType(val interfacePackage: String, val interfaceClassName
   /**
    * Module used to load native libraries.
    */
-  LibraryLoader("com.mapbox.base.module", "LibraryLoader")
+  LibraryLoader("com.mapbox.common.module", "LibraryLoader")
 }
