--- base/src/main/java/com/mapbox/base/module/LibraryLoader.kt
@@ -1,14 +1,0 @@
-package com.mapbox.base.module
-
-import android.content.Context
-
-/**
- * Library loader definition.
- */
-interface LibraryLoader {
-
-  /**
-   * Load native library given a context and library name.
-   */
-  fun load(context: Context, libraryName: String)
-}
