+++ common/gradle.properties
@@ -1,4 +1,4 @@
-IS_ANDROID_PROJECT=true
+IS_ANDROID_PROJECT=false
 POM_ARTIFACT_ID=common
 POM_ARTIFACT_TITLE=Mapbox Common
 POM_DESCRIPTION=Artifact that provides Mapbox module and plugin contracts
