+++ common/src/main/java/com/mapbox/common/module/LibraryLoader.kt
@@ -1,0 +1,12 @@
+package com.mapbox.common.module
+
+/**
+ * Library loader definition.
+ */
+interface LibraryLoader {
+
+  /**
+   * Load native library given a library name.
+   */
+  fun load(libraryName: String)
+}
